Completed migration to Scala 2.13

+++ oscar-cbls/src/main/scala/oscar/cbls/lib/search/combinators/Util.scala
@@ -14,8 +14,8 @@ trait UtilityCombinators{
    * WARNING: do not use this inside an AndThen,
    *          since the objective function is instrumented by this combinator, so the statistics will be counter-intuitive
    *
-   * @param a
-   * @param ignoreInitialObj
+   * @param a The neighborhood to profile
+   * @param ignoreInitialObj flag to ignore the initial value of objective function
    */
   def profile(a:Neighborhood,ignoreInitialObj:Boolean = false) = Profile(a,ignoreInitialObj)
 }
@@ -66,8 +66,8 @@ class ShowObjectiveFunction(a: Neighborhood, obj: () => Long, title: String = "O
  * it will add a prefix to all moves sent back by this combinator
  * the only purposes are documentation and debug
  *
- * @param a
- * @param name
+ * @param a the base neighborhood
+ * @param name the name
  */
 class Name(a: Neighborhood, val name: String) extends NeighborhoodCombinator(a) {
   /**
@@ -96,8 +96,8 @@ class Name(a: Neighborhood, val name: String) extends NeighborhoodCombinator(a) 
  * it will add a prefix to all moves sent back by this combinator
  * the only purposes are documentation and debug
  *
- * @param a
- * @param name
+ * @param a The base neighborhood
+ * @param name The combinator's name
  */
 class ChainableName[MoveType <: Move](a: Neighborhood with SupportForAndThenChaining[MoveType], val name: String)
   extends NeighborhoodCombinator(a) with SupportForAndThenChaining[MoveType]{
@@ -159,11 +159,11 @@ class OverrideObjective(a: Neighborhood, overridingObjective: Objective) extends
    * Some neighborhoods are actually jumps, so that they might violate this basic rule however.
    *
    * @param obj the objective function. notice that it is actually a function. if you have an [[oscar.cbls.core.objective.Objective]] there is an implicit conversion available
-   * @param acceptanceCriterion
+   * @param acceptanceCriterion a function to decide the acceptation of a move
    * @return
    */
   override def getMove(obj: Objective, initialObj:Long, acceptanceCriterion: (Long, Long) => Boolean): SearchResult =
-    getMove(overridingObjective, overridingObjective.value,acceptanceCriterion)
+    getMove(overridingObjective, overridingObjective.value, acceptanceCriterion)
 }
 
 /**
@@ -173,8 +173,8 @@ class OverrideObjective(a: Neighborhood, overridingObjective: Objective) extends
  * WARNING: do not use this inside an AndThen,
  *          since the objective function is instrumented by this combinator, so the statistics will be counter-intuitive
  *
- * @param a
- * @param ignoreInitialObj
+ * @param a The base neighborhood
+ * @param ignoreInitialObj a flag to ignore or not the initial value of objective function
  */
 case class Profile(a:Neighborhood,ignoreInitialObj:Boolean = false) extends NeighborhoodCombinator(a){
 
@@ -205,7 +205,7 @@ case class Profile(a:Neighborhood,ignoreInitialObj:Boolean = false) extends Neig
    * Some neighborhoods are actually jumps, so that they might violate this basic rule however.
    *
    * @param obj the objective function. notice that it is actually a function. if you have an [[oscar.cbls.core.objective.Objective]] there is an implicit conversion available
-   * @param acceptanceCriterion
+   * @param acceptanceCriterion a function to decide the acceptation of a move
    * @return
    */
   override def getMove(obj: Objective, initialObj:Long, acceptanceCriterion: (Long, Long) => Boolean): SearchResult = {
@@ -298,8 +298,8 @@ class Timeout(a:Neighborhood, maxDurationMilliSeconds:Long) extends Neighborhood
 
     if (System.currentTimeMillis() >= deadline) {
 
-      val hours = (maxDurationMilliSeconds / (1000*60*60)).floor.toInt
-      val minutes = (maxDurationMilliSeconds / (1000 * 60)).floor.toInt % 60
+      val hours = (maxDurationMilliSeconds.toFloat / (1000*60*60)).floor.toInt
+      val minutes = (maxDurationMilliSeconds.toFloat / (1000 * 60)).floor.toInt % 60
       val seconds:Double = (maxDurationMilliSeconds / 1000.0) % 60
       println(s"Timeout of $hours:$minutes:$seconds")
       NoMoveFound
